requirements.txt
PyPDF2
lxml
python-docx
pydictionary